refactor(client): rename Client -> LegacyClient (verbatim)
git mv client.ts legacyClient.ts + class rename. Zero body changes.
Importers updated to `LegacyClient as Client` so behavior is identical
and the rename shows as a true file rename in history.
